+++ src/main/java/yqloss/yqlossclientmixinkt/impl/option/TweaksOptionsImpl.kt
@@ -17,10 +17,20 @@ class TweaksOptionsImpl :
 
     @Switch(
         name = "Enable Instant Aim",
-        description = "Set prevRotationYawHead and rotationYawHead of EntityPlayerSP to prevRotationYaw and rotationYaw",
+        description = "Set prevRotationYawHead and rotationYawHead of EntityPlayerSP to prevRotationYaw and rotationYaw.",
         size = 2,
     )
     var enableInstantAimOption = false
 
+    @Switch(
+        name = "Disable Pearl Click-On-Block Packet",
+        description =
+            "Cancel the first C08PacketPlayerBlockPlacement packet when throwing an ender pearl while aiming at a block. " +
+                "This fixes not being able to throw pearls while aiming at a block on public islands in Hypixel SkyBlock.",
+        size = 2,
+    )
+    var disablePearlClickBlockOption = false
+
     override val enableInstantAim by ::enableInstantAimOption
+    override val disablePearlClickBlock by ::disablePearlClickBlockOption
 }