+++ src/main/java/yqloss/yqlossclientmixinkt/module/tweaks/TweaksOptions.kt
@@ -4,4 +4,5 @@ import yqloss.yqlossclientmixinkt.module.YCModuleOptions
 
 interface TweaksOptions : YCModuleOptions {
     val enableInstantAim: Boolean
+    val disablePearlClickBlock: Boolean
 }