+++ CHANGELOG.md
@@ -8,3 +8,5 @@
 * feat Tweaks: Instant Aim
 
 ## 0.2.0 (Under Development)
+
+* feat Tweaks: Cancel Pearl Click-On-Block Packet